+++ R/dbi-connection.R
@@ -148,13 +148,13 @@ setMethod("dbQuoteIdentifier", c("OdbcConnection", "character"),
       return(DBI::SQL(character()))
     }
     if (any(is.na(x))) {
-      stop("Cannot pass NA to dbQuoteIdentifier()", call. = FALSE)
+      cli::cli_abort("{.arg x} can't be {.code NA}.")
     }
     if (nzchar(conn@quote)) {
       x <- gsub(conn@quote, paste0(conn@quote, conn@quote), x, fixed = TRUE)
     }
     nms <- names(x)
-    res <- DBI::SQL(paste(conn@quote, encodeString(x), conn@quote, sep = ""))
+    res <- DBI::SQL(paste(conn@quote, x, conn@quote, sep = ""))
     names(res) <- nms
     res
   }
@@ -191,6 +191,11 @@ setMethod("dbListTables", "OdbcConnection",
            table_name = NULL,
            table_type = NULL,
            ...) {
+    check_string(catalog_name, allow_null = TRUE)
+    check_string(schema_name, allow_null = TRUE)
+    check_string(table_name, allow_null = TRUE)
+    check_string(table_type, allow_null = TRUE)
+
     tables <- odbcConnectionTables(
       conn,
       name = table_name,
@@ -200,29 +205,6 @@ setMethod("dbListTables", "OdbcConnection",
       exact = TRUE
     )
     tables[["table_name"]]
-  }
-)
-
-#' @rdname dbListTables-OdbcConnection-method
-#' @inheritParams DBI::dbListFields
-#' @param column_name The name of the column to return, the default returns all columns.
-#' @export
-setMethod("dbListFields", c("OdbcConnection", "character"),
-  function(conn,
-           name,
-           catalog_name = NULL,
-           schema_name = NULL,
-           column_name = NULL,
-           ...) {
-    cols <- odbcConnectionColumns_(
-      conn,
-      name = name,
-      catalog_name = catalog_name,
-      schema_name = schema_name,
-      column_name = column_name,
-      exact = TRUE
-    )
-    cols[["name"]]
   }
 )