+++ R/dbi-result.R
@@ -63,7 +63,8 @@ setMethod("dbClearResult", "OdbcResult",
 #' @export
 setMethod("dbFetch", "OdbcResult",
   function(res, n = -1, ...) {
-    n <- check_n(n)
+    check_number_whole(n, min = -1, allow_infinite = TRUE)
+    if (is.infinite(n)) n <- -1
     result_fetch(res@ptr, n)
   }
 )
@@ -92,7 +93,7 @@ setMethod("dbIsValid", "OdbcResult",
 setMethod("dbGetStatement", "OdbcResult",
   function(res, ...) {
     if (!dbIsValid(res)) {
-      stop("Result already cleared", call. = FALSE)
+      cli::cli_abort("Result already cleared.")
     }
     res@statement
   }