+++ R/driver-hana.R
@@ -14,6 +14,9 @@ setMethod("sqlCreateTable", "HDB",
            temporary = FALSE,
            ...,
            field.types = NULL) {
+    check_bool(temporary)
+    check_row.names(row.names)
+    check_field.types(field.types)
     table <- dbQuoteIdentifier(con, table)
     fields <- createFields(con, fields, field.types, row.names)